requirements.txt
torch>=2.1.0
torchvision>=0.16.0
flwr>=1.7.0
opacus>=1.4.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
Pillow>=10.0.0
pyyaml>=6.0
kagglehub>=0.3.0
tqdm>=4.65.0
grad-cam>=1.4.8
wandb>=0.15.0
jupyter>=1.0.0
pytest>=7.4.0
